bug: fix cleanup metadata task (#39)
* bug: fix issue with attempting to delete same metadata multiple times

* bug: issue with metadata cleanup

Fixed two issues with metadata cleanup:

Possible error trying to delete metadata without a record:
Added a check for metadata id

Only handling each resource once for the first extractor:
Moved extraction record deletion outside of extractor specific
logic, to prevent metadata records only being deleted for one
extractor

* bug: cannot get_in_or_equal empty array

+++ classes/task/cleanup_metadata_task.php
@@ -102,23 +102,34 @@ class cleanup_metadata_task extends scheduled_task {
         if (empty($enabledplugins)) {
             mtrace('tool_metadata: No enabled metadata subplugins, metadata cleanup skipped.');
         } else {
+            $processedresourcehashes = [];
+
             foreach ($enabledplugins as $plugin) {
                 $deletecount = 0;
                 $extractor = api::get_extractor($plugin);
                 $resourcehashes = $this->get_deleted_resourcehashes($extractor);
+                // Remove duplicates.
+                $resourcehashes = array_unique($resourcehashes);
                 if (!empty($resourcehashes)) {
                     foreach ($resourcehashes as $resourcehash) {
                         $metadata = $extractor->get_metadata($resourcehash);
-                        $metadata->delete();
-                        $deletecount++;
+                        if (!empty($metadata) && !empty($metadata->id)) {
+                            $metadata->delete();
+                            $deletecount++;
+                            $processedresourcehashes[] = $resourcehash;
+                        }
                     }
-                    // Delete the extraction records associated with resourcehashes too.
-                    list($insql, $inparams) = $DB->get_in_or_equal($resourcehashes);
-                    $DB->delete_records_select(extraction::TABLE, 'resourcehash ' . $insql, $inparams);
                     mtrace("tool_metadata: Deleted $deletecount metadata records for metadataextractor_$plugin");
                 } else {
                     mtrace("tool_metadata: No metadata to cleanup for metadataextractor_$plugin");
                 }
+            }
+
+            if (!empty($processedresourcehashes)) {
+                // Delete the extraction records associated with processed resourcehashes too.
+                $processedresourcehashes = array_unique($processedresourcehashes);
+                list($insql, $inparams) = $DB->get_in_or_equal($processedresourcehashes);
+                $DB->delete_records_select(extraction::TABLE, 'resourcehash ' . $insql, $inparams);
             }
         }
     }
